+++ include/mruby_integration/models/font.hpp
@@ -5,5 +5,5 @@
 
 extern RClass *Font_class;
 
-void setup_Font(mrb_state*, mrb_value, Font*, int, int, int);
+void setup_Font(mrb_state*, mrb_value, Font*, int, int, int, Texture2D*);
 void append_models_Font(mrb_state*);