+++ mrb_doc/textures.rb
@@ -60,3 +60,22 @@ def fade(colour)
   # src/mruby_integration/textures.cpp
   Colour.new
 end
+
+# Generates mipmaps for the {Texture2D}
+# @param texture [Texture2D]
+# @return [nil]
+def generate_texture_mipmaps(texture)
+  # mrb_gen_texture_mipmaps
+  # src/mruby_integration/textures.cpp
+  nil
+end
+
+# Sets the filtering for the {Texture2D}
+# @param texture [Texture2D]
+# @param filter [Integer] What sort of filtering to apply, valid options are: {TEXTURE_FILTER_POINT}, {TEXTURE_FILTER_BILINEAR}, {TEXTURE_FILTER_TRILINEAR}, {TEXTURE_FILTER_ANISOTROPIC_4X}, {TEXTURE_FILTER_ANISOTROPIC_8X}, or {TEXTURE_FILTER_ANISOTROPIC_16X}
+# @return [nil]
+def set_texture_filter(texture, filter)
+  # mrb_set_texture_filter
+  # src/mruby_integration/textures.cpp
+  nil
+end