+++ src/mruby_integration/text.cpp
@@ -15,7 +15,10 @@ mrb_value mrb_load_font(mrb_state *mrb, mrb_value) {
 
   mrb_value obj = mrb_obj_value(Data_Wrap_Struct(mrb, Font_class, &Font_type, font));
 
-  setup_Font(mrb, obj, font, font->baseSize, font->glyphCount, font->glyphPadding);
+  setup_Font(mrb, obj, font, font->baseSize, font->glyphCount, font->glyphPadding, &font->texture);
+
+  add_parent(font, "Font");
+  add_owned_object(&font->texture);
 
   return obj;
 }
@@ -31,7 +34,10 @@ mrb_value mrb_load_font_ex(mrb_state *mrb, mrb_value) {
 
   mrb_value obj = mrb_obj_value(Data_Wrap_Struct(mrb, Font_class, &Font_type, font));
 
-  setup_Font(mrb, obj, font, font->baseSize, font->glyphCount, font->glyphPadding);
+  setup_Font(mrb, obj, font, font->baseSize, font->glyphCount, font->glyphPadding, &font->texture);
+
+  add_parent(font, "Font");
+  add_owned_object(&font->texture);
 
   return obj;
 }